Add MX-emulation ladder-selection pipeline
Adds the MX/FP8 quantization + bound-eligibility ladder selector on top
of the existing PLASMA tile-Cholesky engine:

- compute/mx_emulation.{h,cpp}, mx_apply.h, ladder_selection.h -- the
  bound-eligibility selector, which picks the cheapest per-tile format
  whose error bound still holds, instead of the fixed epsilon-ratio
  heuristic. Adds three modes (MX_BOUND_LADDER / LADDER=):
  ladder_ieee, ladder_mx, ladder_mx_mxfp16. The original heuristic is
  untouched and still reachable as LADDER=baseline.
- MX shared-scale tiers (e2m1 / mx_e4m3 / mx_fp16, 32-element row-vector
  scale groups) plus 6 MX metadata fields on MixedPrecisionTile.
- compute/dist_tile_copy.h, mx_tile_copy.h -- deduceMemcpyKind: the two
  pdpotrf_gpu_*.cpp paths hardcoded cudaMemcpyHostToDevice/DeviceToHost;
  they now inspect the pointer via cudaPointerGetAttributes so device and
  managed memory copy correctly, plus a requantizeTileHost hook after
  each diagonal-tile writeback.
- run_ladder.sh, run_ladder_ftz_tilemap_nb4096.sh,
  tools/extract_tilemap.py -- sweep and tile-allocation drivers.
- data/ input handling: paths are repo-relative and overridable rather
  than absolute, with data/README.md documenting the expected format
  (raw row-major NxN float64). No .bin files are tracked.
- Build/portability: LAPACKE detection, std::byte->uint8_t, min<> cast,
  and CMAKE_CUDA_ARCHITECTURES = "80-real;90-real;90-virtual" -- native
  SASS for A100 and Hopper, plus compute_90 PTX so newer architectures
  (Blackwell and beyond, untested) run via driver JIT rather than
  failing with cudaErrorNoKernelImageForDevice.

Verified on A100: stock example_dpotrf still passes its own
factorization check, and the ladder runs clean end-to-end. Ladder
numerics were checked at eps=1e-5/gu on the 32k matrix across all three
modes (ladder_mx_mxfp16 rel=2.1901e-05, ladder_ieee rel=2.1932e-06,
baseline rel=0.00029708).

+++ compute/castType.cu
@@ -6,7 +6,7 @@
 #include <cstdio>
 
 #include "mixed_precision.h"
-
+#include <algorithm>
 template <typename Src, typename Dst>
 __global__ void castKernel(int m, int n, const void *source, int ldSource,
                            void *target, int ldTarget) {
@@ -26,7 +26,7 @@ void *castToBuffer(int m, int n, const void *source, cudaDataType sourceType,
   if (targetType == sourceType) {
     return const_cast<void *>(source);
   }
-  dim3 block{std::min<unsigned int>(256, m * n)};
+  dim3 block{std::min<unsigned int>(256, static_cast<unsigned int>(m * n))};
   dim3 grid{(m * n - 1) / block.x + 1};
   switch (sourceType) {
     case CUDA_R_64F: {
